fix(editor): 트리거 액션 리뷰 반영
레거시 정보 공개 액션도 신규 정보 공개 필드와 필수값 검증을 사용하도록 맞추고, 커진 ActionListEditor의 필드 렌더링 책임을 별도 컴포넌트와 helper로 분리했습니다.

Constraint: PR #430 CodeRabbit review thread 2건만 처리하고 runtime engine 계약과 다른 에디터 흐름은 변경하지 않았습니다.

Tested: pnpm --dir apps/web test -- ActionListEditor.test.tsx

Tested: pnpm --dir apps/web typecheck

Tested: pnpm --dir apps/web exec eslint src/features/editor/components/design/ActionListEditor.tsx src/features/editor/components/design/PresentationCueFields.tsx src/features/editor/components/design/InformationActionFields.tsx src/features/editor/components/design/BroadcastActionFields.tsx src/features/editor/components/design/actionFieldHelpers.ts src/features/editor/components/design/__tests__/ActionListEditor.test.tsx
